fix authentication

+++ frontend-react/src/features/settings/SettingsPage.tsx
@@ -1,22 +1,35 @@
 import { Button } from '@/shared/components/ui/button'
 import { BiLogOut } from 'react-icons/bi'
 import { FiTrash2 } from 'react-icons/fi'
-import { useState } from 'react'
-import { api } from '@/lib/api-client'
+import { useContext, useState } from 'react'
 import { useUserDelete } from '@/shared/hooks/users'
 import { useAccount } from '@/shared/hooks/account'
 import ConfirmationModal from '../../shared/components/ui/custom/ConfirmationModal.tsx'
+import { useNavigate } from 'react-router'
+import { AuthContext } from '@/providers/AuthProvider.tsx'
 
 const SettingsPage = () => {
+  const navigate = useNavigate()
   const [showLogoutDialog, setShowLogoutDialog] = useState(false)
   const [showDeleteDialog, setShowDeleteDialog] = useState(false)
   const { mutate: deleteUser } = useUserDelete()
   const { user_id } = useAccount()
+  const { logout } = useContext(AuthContext)
 
   const handleDeleteUser = () => {
+    if (!user_id) return
     deleteUser(user_id, {
-      onSuccess: () => api.logout(),
+      onSuccess: () => {
+        logout()
+        navigate('/auth/login')
+      },
     })
+  }
+
+  const handleLogOut = () => {
+    setShowLogoutDialog(false)
+    logout()
+    navigate('/auth/login')
   }
 
   return (
@@ -24,7 +37,7 @@ const SettingsPage = () => {
       <ConfirmationModal
         open={showLogoutDialog}
         onOpenChange={setShowLogoutDialog}
-        onConfirm={() => api.logout()}
+        onConfirm={handleLogOut}
         title="Se déconnecter"
         description="Êtes-vous sûr de vouloir vous déconnecter ?"
         confirmText="Confirmer la déconnexion"
